# todo-py-fastapi/app/models.py
# generated by fastapi-codegen:
#   filename:  openapi3-todo.yaml
#   timestamp: 2022-10-26T14:53:57+00:00

from __future__ import annotations

from enum import Enum
from typing import Optional

from pydantic import BaseModel, Field


class State(Enum):
    ACTIVE = 'ACTIVE'
    COMPLETED = 'COMPLETED'
    DELETED = 'DELETED'


class Todo(BaseModel):
    uuid: str = Field(..., description='Unique identifier of the todo')
    ownerUuid: str = Field(
        ..., description='Unique identifier of the owner of the todo'
    )
    title: str = Field(..., description='Title/short summary of the todo')
    state: State = Field(..., description='State of the todo')
    description: Optional[str] = Field(
        None, description='The lengthy description of this todo'
    )


class Error(BaseModel):
    statusCode: Optional[float] = Field(None, description='Status code of the response')
    error: Optional[str] = Field(None, description='Name of the error')
    message: Optional[str] = Field(None, description='Explanation of the error')